creating shared folder for shared vue files
