+++ resources/js/Pages/main.vue
@@ -2,114 +2,22 @@
     import { library } from '@fortawesome/fontawesome-svg-core';
     import { faBars } from '@fortawesome/free-solid-svg-icons';
     import { faXmark } from '@fortawesome/free-solid-svg-icons';
-    import { FontAwesomeIcon } from '@fortawesome/vue-fontawesome';
-
-    import {
-        ref
-    } from 'vue'
+    import navbar from '../Shared/navbar.vue';
+    import navbarMobile from '../Shared/navbarMobile.vue';
+    import {ref} from 'vue';
 
     library.add(faBars, faXmark);
 
-    const navigation = [{
-            name: 'Company',
-            href: '#'
-        },
-        {
-            name: 'Blog',
-            href: '#'
-        },
-        {
-            name: 'Marketplace',
-            href: '#'
-        },
-        {
-            name: 'Features',
-            href: '#'
-        },
-    ]
-
     const mobileMenuOpen = ref(false)
+
 </script>
 <template>
     <div class="bg-white">
         <header class="absolute inset-x-0 top-0 z-50">
-            <nav class="flex items-center justify-between p-6 lg:px-8" aria-label="Global">
+            
+            <navbar></navbar>
 
-                <div class="flex lg:flex-1">
-                    <a href="#" class="-m-1.5 p-1.5">
-                        <img class="h-10 w-auto" src="build/assets/images/keyvet.png" alt="" />
-                    </a>
-                </div>
-                <!--start humburger icon -->
-                <div class="flex lg:hidden">
-                    <button type="button"
-                        class="-m-2.5 inline-flex items-center justify-center rounded-md p-2.5 text-gray-700"
-                        @click.prevent="mobileMenuOpen = true">
-
-                        <font-awesome-icon icon="fa-solid fa-bars" class="text-3xl text-blue-500"></font-awesome-icon>
-
-                    </button>
-                </div>
-                <!--end humburger icon -->
-
-                <div class="hidden lg:flex lg:gap-x-12">
-                    <a v-for="item in navigation" :key="item.name" :href="item.href"
-                        class="text-sm font-semibold leading-6 text-gray-900 hover:text-blue-500">{{ item . name }}</a>
-                </div>
-
-
-                <div class="hidden lg:flex lg:flex-1 lg:justify-end">
-                    <a href="#" class="text-sm p-2 font-semibold leading-6 text-gray-900 hover:text-blue-500">
-                        Register
-                    </a>
-                    <a href="#"
-                        class="text-sm leading-8 rounded-md bg-blue-400 px-3 py-1 font-semibold text-white shadow-sm hover:bg-blue-500 focus-visible:outline focus-visible:outline-2 focus-visible:outline-offset-2 focus-visible:outline-blue-600">
-                        Log in
-                        <span aria-hidden="true">&rarr;
-                        </span>
-                    </a>
-                </div>
-            </nav>
-            <div class="lg:hidden" @click="mobileMenuOpen = false">
-                <Transition
-                enter-from-class="transition duration-200 opacity-0"
-                enter-to-class="transition duration-200 opacity-100"
-                leave-from-class="transition duration-200 opacity-100"
-                leave-to-class="transition duration-200 opacity-0"
-                >
-                    <div v-if="mobileMenuOpen"
-                        class="fixed inset-y-0 right-0 z-50 w-full overflow-y-auto bg-white px-6 py-6 sm:max-w-sm sm:ring-1 sm:ring-gray-900/10">
-                        <div class="flex items-center justify-between">
-                            <a href="#" class="-m-1.5 p-1.5">
-                                <img class="h-8 w-auto" src="build/assets/images/keyvet.png" alt="" />
-                            </a>
-                            <button @click="mobileMenuOpen = false" class="-m-2.5 rounded-md p-2.5 text-gray-700">
-                                
-                                <font-awesome-icon icon="fa-solid fa-xmark" class="text-3xl text-blue-500"></font-awesome-icon>
-
-                            </button>
-                        </div>
-                        <div class="mt-6 flow-root">
-                            <div class="-my-6 divide-y divide-gray-500/10">
-                                <div class="space-y-2 py-6">
-                                    <a v-for="item in navigation" :key="item.name" :href="item.href"
-                                        class="-mx-3 block rounded-lg px-3 py-2 text-base font-semibold leading-7 text-gray-900 hover:bg-gray-50">{{ item . name }}</a>
-                                </div>
-                                <div class="py-6">
-                                    <a href="#"
-                                        class="-mx-3 block rounded-lg px-3 py-2.5 text-base font-semibold leading-7 text-gray-900 hover:bg-gray-50">Log
-                                        in
-                                    </a>
-                                    <a href="#"
-                                        class="-mx-3 block rounded-lg px-3 py-2.5 text-base font-semibold leading-7 text-gray-900 hover:bg-gray-50">
-                                        Register
-                                    </a>
-                                </div>
-                            </div>
-                        </div>
-                    </div>
-                </Transition>
-            </div>
+            <navbarMobile :mobileMenuOpen="mobileMenuOpen"></navbarMobile>
         </header>
 
         <div class="relative isolate px-6 pt-14 lg:px-8">
